+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+v0.1.10
+    * Replaced `--no-*-selectors` by `no-*`.
+    * Introduced `--column-size` option.
+    * Fixed bugs.
 v0.1.9
-    Introduced query command.
+    * Introduced query command.